Added validation for Update/Delete + Owner field autopopulated

+++ src/main/java/com/example/demo/Service/ProductService.java
@@ -3,57 +3,47 @@ package com.example.demo.Service;
 import com.example.demo.DTO.ProductDTO;
 import com.example.demo.DTO.ProductResponse;
 import com.example.demo.DTO.ReviewDTO;
+import com.example.demo.DTO.SingleProductDTO;
 import com.example.demo.Entity.Product;
 import com.example.demo.Entity.Review;
 import com.example.demo.Entity.User;
 import com.example.demo.Repository.ProductRepository;
-import com.example.demo.Repository.ReviewRepository;
 import com.example.demo.Repository.UserRepository;
-import com.example.demo.Security.JWTGenerator;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.nio.file.AccessDeniedException;
 import java.util.List;
 import java.util.stream.Collectors;
 
 @Service
 public class ProductService {
 
-    ProductRepository productRepository;
-//    ReviewRepository reviewRepository;
-    UserRepository userRepository;
-    JWTGenerator jwtGenerator;
+    private final ProductRepository productRepository;
+    private final UserRepository userRepository;
 
     @Autowired
-    public ProductService(ProductRepository productRepository,
-                          UserRepository userRepository
-//                          ReviewRepository reviewRepository
-                          ,JWTGenerator jwtGenerator
-    ){
+    public ProductService(ProductRepository productRepository, UserRepository userRepository){
         this.productRepository = productRepository;
-//        this.reviewRepository = reviewRepository;
         this.userRepository = userRepository;
-        this.jwtGenerator = jwtGenerator;
     }
 
-    public ProductDTO createProduct(ProductDTO productDTO){
+    public ProductDTO createProduct(ProductDTO productDTO, String username){
         Product product = new Product();
         product.setName(productDTO.getName());
         product.setDescription(productDTO.getDescription());
         product.setPrice(productDTO.getPrice());
 
         // Associate it with the Owner
-        User owner = userRepository.findById(productDTO.getOwnerId()).orElseThrow(()->new RuntimeException("User not found:("));
+        User owner = userRepository.findByUsername(username).orElseThrow(()->new RuntimeException("User not found:("));
         product.setOwner(owner);
 
         Product newProduct = productRepository.save(product);
-//        newProduct.setOwner();
-        // Populate Owner Field
-//        productResponse.setOwnerId(newProduct.getOwner().getId());
 
         return mapToDTO(newProduct);
     }
@@ -76,14 +66,15 @@ public class ProductService {
         return productResponse;
     }
 
-    public ProductDTO getProductById(Long id){
+    public SingleProductDTO getProductById(Long id){
         Product product = productRepository.findById(id).orElseThrow(()->new RuntimeException("Product not found:("));
-//        List<Review> reviews = reviewRepository.findByProductId(id);
-//        product.setReviews(reviews);
-        return mapToDTO(product);
+        return mapProductToDTO(product);
     }
 
-    public ProductDTO updateProduct(Long id, ProductDTO productDTO){
+    @Transactional
+    public ProductDTO updateProduct(Long id, ProductDTO productDTO) throws AccessDeniedException {
+        validateProductOwner(id);
+
         Product product = productRepository.findById(id).orElseThrow(()->new RuntimeException("Product not found:("));
 
         product.setName(productDTO.getName());
@@ -94,10 +85,23 @@ public class ProductService {
         return mapToDTO(productToUpdate);
     }
 
-    public String deleteProduct(Long id){
+    @Transactional
+    public String deleteProduct(Long id) throws AccessDeniedException {
+        validateProductOwner(id);
+
         Product product = productRepository.findById(id).orElseThrow(()->new RuntimeException("Product not found:("));
         productRepository.delete(product);
         return "Success!";
+    }
+
+    private void validateProductOwner(Long productId) throws AccessDeniedException {
+        String currentUsername = SecurityContextHolder.getContext().getAuthentication().getName();
+        User currentUser = userRepository.findByUsername(currentUsername).orElseThrow(()-> new RuntimeException("User not found:("));
+        Product product  = productRepository.findById(productId).orElseThrow(() -> new RuntimeException("No Product here:("));
+
+        if(!product.getOwner().equals(currentUser)){
+            throw new AccessDeniedException("Unauthorized!");
+        }
     }
 
     private ProductDTO mapToDTO(Product product){
@@ -106,9 +110,28 @@ public class ProductService {
         productDTO.setName(product.getName());
         productDTO.setDescription(product.getDescription());
         productDTO.setPrice(product.getPrice());
-//        productDTO.setReviews(product.getReviews());
         productDTO.setOwnerId(product.getOwner().getId());
         return productDTO;
+    }
+
+    private SingleProductDTO mapProductToDTO(Product product){
+        SingleProductDTO productDTO = new SingleProductDTO();
+        productDTO.setId(product.getId());
+        productDTO.setName(product.getName());
+        productDTO.setDescription(product.getDescription());
+        productDTO.setPrice(product.getPrice());
+        productDTO.setReviews(product.getReviews().stream().map(this::mapReviewToDTO).collect(Collectors.toList()));
+        productDTO.setOwnerId(product.getOwner().getId());
+        return productDTO;
+    }
+
+    private ReviewDTO mapReviewToDTO(Review review){
+        ReviewDTO reviewDTO = new ReviewDTO();
+        reviewDTO.setId(review.getId());
+        reviewDTO.setContent(review.getContent());
+        reviewDTO.setRating(review.getRating());
+        reviewDTO.setOwnerId(review.getOwner().getId());
+        return reviewDTO;
     }
 
     private Product mapToEntity(ProductDTO productDTO){
@@ -118,5 +141,4 @@ public class ProductService {
         product.setPrice(productDTO.getPrice());
         return product;
     }
-
 }
